third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-stdDeviation-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.stdDeviationX.baseVal is 0
PASS dropShadowElement.stdDeviationY.baseVal is 0
PASS dropShadowElement.stdDeviationX.baseVal is 10
PASS dropShadowElement.stdDeviationY.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
